+++ mail/mail-common/src/models/message.rs
@@ -1343,7 +1343,9 @@ impl Message {
             return Err(AppError::MessageMissing(message_id));
         };
 
-        let reply_actions = if message.to_list.len() + message.cc_list.len() > 1 {
+        let reply_actions = if !message.can_reply() {
+            vec![]
+        } else if message.to_list.len() + message.cc_list.len() > 1 {
             ReplyAction::all()
         } else {
             ReplyAction::single_address()